refactor(mnemonic): port MnemonicCode to commonMain as an object
